requirements.txt
asgiref==3.7.2
black==23.9.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
distlib==0.3.7
Django==4.1.5
django-appconf==1.0.5
django-cors-headers==4.2.0
django-filter==23.2
django-imagekit==4.1.0
django-import-export==3.3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
filelock==3.12.4
flake8==6.1.0
flake8-broken-line==1.0.0
flake8-isort==6.1.0
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
google-api-core==2.12.0
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
gunicorn==21.2.0
httplib2==0.22.0
identify==2.5.28
idna==3.4
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
pep8-naming==0.13.3
pilkit==2.0
Pillow==9.5.0
platformdirs==3.10.0
pre-commit==3.4.0
protobuf==4.24.4
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.0
python-dotenv-run==0.1.4
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruff==0.0.290
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
virtualenv==20.24.5
